Load map from file.

+++ core/src/main/kotlin/gabek/sm2/systems/graphics/HealthRenderSystem.kt
@@ -9,8 +9,8 @@ import com.badlogic.gdx.math.Rectangle
 import com.github.salomonbrys.kodein.Kodein
 import com.github.salomonbrys.kodein.instance
 import gabek.sm2.assets.Assets
-import gabek.sm2.components.character.HealthCom
 import gabek.sm2.components.TranslationCom
+import gabek.sm2.components.character.HealthCom
 import gabek.sm2.components.graphics.HealthDisplayCom
 import gabek.sm2.world.RenderManager
 
@@ -22,13 +22,13 @@ class HealthRenderSystem(kodein: Kodein) : BaseEntitySystem(
                 HealthCom::class.java,
                 HealthDisplayCom::class.java,
                 TranslationCom::class.java)
-), RenderManager.BatchSystem{
+), RenderManager.BatchSystem {
     private val assets: Assets = kodein.instance()
 
     private val rect = assets.retrieveRegion("actors:rect")
 
-    private val background = Color(89/255f, 86/255f, 82/255f, 1f)
-    private val foreground = Color(106/255f, 190/255f, 48/255f, 1f)
+    private val background = Color(89 / 255f, 86 / 255f, 82 / 255f, 1f)
+    private val foreground = Color(106 / 255f, 190 / 255f, 48 / 255f, 1f)
 
     private val temp = Rectangle()
 
@@ -37,10 +37,10 @@ class HealthRenderSystem(kodein: Kodein) : BaseEntitySystem(
     private lateinit var healthDisplayMapper: ComponentMapper<HealthDisplayCom>
 
     override fun processSystem() {}
-    override fun render(batch: SpriteBatch, culling: Rectangle, progress: Float){
+    override fun render(batch: SpriteBatch, culling: Rectangle, progress: Float) {
         val entities = entityIds
 
-        for(i in 0 until entities.size()){
+        for (i in 0 until entities.size()) {
             val entity = entities[i]
 
             val trans = transMapper[entity]
@@ -53,12 +53,12 @@ class HealthRenderSystem(kodein: Kodein) : BaseEntitySystem(
             temp.x = trans.lerpX(progress) + healthDisplay.offsetX - temp.width / 2f
             temp.y = trans.lerpY(progress) + healthDisplay.offsetY - temp.height / 2f
 
-            if(culling.overlaps(temp)) {
+            if (culling.overlaps(temp)) {
                 batch.color = background
                 batch.draw(rect, temp.x, temp.y, temp.width, temp.height)
 
                 batch.color = foreground
-                batch.draw(rect, temp.x, temp.y, temp.width * (health.healthPoints / health.maximumHealth.toFloat()), temp.height)
+                batch.draw(rect, temp.x, temp.y, temp.width * (health.healthPoints / health.maximumHealth), temp.height)
 
                 batch.color = Color.WHITE
             }
